@@ -1,16 +1,16 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <Platforms>x64</Platforms>
     <Company>Den Delimarsky</Company>
     <Authors>Den Delimarsky</Authors>
-    <Version>0.0.3</Version>
+    <Version>0.0.4</Version>
     <ApplicationIcon>piglet.ico</ApplicationIcon>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <Copyright>Copyright © 2021 by Den Delimarsky. All rights reserved.</Copyright>
+    <Copyright>Copyright © 2024 by Den Delimarsky. All rights reserved.</Copyright>
     <PackageProjectUrl>https://deck.surf</PackageProjectUrl>
     <PackageIcon>logo.png</PackageIcon>
     <PackageIconUrl />
@@ -22,9 +22,9 @@
     <RepositoryUrl>https://github.com/dend/decksurf-sdk</RepositoryUrl>
     <PackageTags>streamdeck,sdk,decksurf,hid,buttons</PackageTags>
     <NeutralLanguage>en-US</NeutralLanguage>
-    <AssemblyVersion>0.0.3.0</AssemblyVersion>
-    <FileVersion>0.0.3.0</FileVersion>
-    <Description>Open-source package to manager your Stream Deck device from a .NET project.</Description>
+    <AssemblyVersion>0.0.4.0</AssemblyVersion>
+    <FileVersion>0.0.4.0</FileVersion>
+    <Description>Open-source SDK to manage your Stream Deck device from .NET.</Description>
     <RepositoryType>git</RepositoryType>
   </PropertyGroup>
 
@@ -62,7 +62,7 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="System.Drawing.Common" Version="6.0.0-preview.4.21253.7" />
+    <PackageReference Include="System.Drawing.Common" Version="9.0.0" />
   </ItemGroup>
 
 </Project>
