@@ -4,10 +4,10 @@
 
 using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Threading;
-using DeckSurf.SDK.Core;
 using DeckSurf.SDK.Util;
 using HidSharp;
 
@@ -18,11 +18,11 @@ namespace DeckSurf.SDK.Models
     /// </summary>
     public abstract class ConnectedDevice
     {
-        private const int ButtonPressHeaderOffset = 4;
-
         private static readonly int ImageReportLength = 1024;
         private static readonly int ImageReportHeaderLength = 8;
+        private static readonly int ImageReportScreenHeaderLength = 16;
         private static readonly int ImageReportPayloadLength = ImageReportLength - ImageReportHeaderLength;
+        private static readonly int ImageReportScreenPayloadLength = ImageReportLength - ImageReportScreenHeaderLength;
 
         private byte[] keyPressBuffer = new byte[1024];
 
@@ -40,23 +40,12 @@ namespace DeckSurf.SDK.Models
         /// <param name="pid">Product ID.</param>
         /// <param name="path">Path to the USB HID device.</param>
         /// <param name="name">Name of the USB HID device.</param>
-        /// <param name="model">Stream Deck model.</param>
-        public ConnectedDevice(int vid, int pid, string path, string name, DeviceModel model)
+        public ConnectedDevice(int vid, int pid, string path, string name)
         {
             this.VId = vid;
-            this.PId = pid;
             this.Path = path;
             this.Name = name;
-            this.Model = model;
-            this.UnderlyingDevice = DeviceList.Local.GetHidDeviceOrNull(this.VId, this.PId);
-
-            this.ButtonCount = model switch
-            {
-                DeviceModel.XL => DeviceConstants.XLButtonCount,
-                DeviceModel.MINI => DeviceConstants.MiniButtonCount,
-                DeviceModel.ORIGINAL or DeviceModel.ORIGINAL_V2 => DeviceConstants.OriginalButtonCount,
-                _ => 0,
-            };
+            this.UnderlyingDevice = DeviceList.Local.GetHidDeviceOrNull(vid, pid);
         }
 
         /// <summary>
@@ -74,32 +63,81 @@ namespace DeckSurf.SDK.Models
         /// <summary>
         /// Gets the vendor ID.
         /// </summary>
-        public int VId { get; private set; }
-
-        /// <summary>
-        /// Gets the product ID.
-        /// </summary>
-        public int PId { get; private set; }
+        public int VId { get; }
 
         /// <summary>
         /// Gets the USB HID device path.
         /// </summary>
-        public string Path { get; private set; }
+        public string Path { get; }
 
         /// <summary>
         /// Gets the USB HID device name.
         /// </summary>
-        public string Name { get; private set; }
+        public string Name { get; }
 
         /// <summary>
         /// Gets the Stream Deck device model.
         /// </summary>
-        public DeviceModel Model { get; private set; }
+        public abstract DeviceModel Model { get; }
 
         /// <summary>
         /// Gets the number of buttons on the connected Stream Deck device.
         /// </summary>
-        public int ButtonCount { get; }
+        public abstract int ButtonCount { get; }
+
+        /// <summary>
+        /// Gets a value indicating whether the image needs to be flipped before being passed to a button.
+        /// </summary>
+        public abstract bool IsButtonImageFlipRequired { get; }
+
+        /// <summary>
+        /// Gets a value indicating whether the Stream Deck device has a screen in addition to buttons.
+        /// </summary>
+        public abstract bool IsScreenSupported { get; }
+
+        /// <summary>
+        /// Gets a value indicating whether the Stream Deck has knobs.
+        /// </summary>
+        public abstract bool IsKnobSupported { get; }
+
+        /// <summary>
+        /// Gets a value indicating the button resolution for the Stream Deck device.
+        /// </summary>
+        public abstract int ButtonResolution { get; }
+
+        /// <summary>
+        /// Gets a value indicating the number of button columns for a Stream Deck device.
+        /// </summary>
+        public abstract int ButtonColumns { get; }
+
+        /// <summary>
+        /// Gets a value indicating the number of button rows for a Stream Deck device.
+        /// </summary>
+        public abstract int ButtonRows { get; }
+
+        /// <summary>
+        /// Gets screen width for the Stream Deck Plus.
+        /// </summary>
+        /// <remarks>
+        /// Returns -1 if there is no screen supported.
+        /// </remarks>
+        public abstract int ScreenWidth { get; }
+
+        /// <summary>
+        /// Gets screen height for the Stream Deck device that supports a screen.
+        /// </summary>
+        /// <remarks>
+        /// Returns -1 if there is no screen supported.
+        /// </remarks>
+        public abstract int ScreenHeight { get; }
+
+        /// <summary>
+        /// Gets screen width for a segment on the Stream Deck device that supports a screen.
+        /// </summary>
+        /// <remarks>
+        /// Returns -1 if there is no screen supported.
+        /// </remarks>
+        public abstract int ScreenSegmentWidth { get; }
 
         private HidDevice UnderlyingDevice { get; }
 
@@ -108,11 +146,19 @@ namespace DeckSurf.SDK.Models
         /// <summary>
         /// Initialize the device and start reading the input stream.
         /// </summary>
-        public void InitializeDevice()
+        public void StartListening()
         {
             this.UnderlyingInputStream = this.UnderlyingDevice.Open();
             this.UnderlyingInputStream.ReadTimeout = Timeout.Infinite;
             this.UnderlyingInputStream.BeginRead(this.keyPressBuffer, 0, this.keyPressBuffer.Length, this.KeyPressCallback, null);
+        }
+
+        /// <summary>
+        /// Stops listening for events for the specific device.
+        /// </summary>
+        public void StopListening()
+        {
+            this.UnderlyingInputStream.Close();
         }
 
         /// <summary>
@@ -127,13 +173,11 @@ namespace DeckSurf.SDK.Models
         /// <summary>
         /// Clear the contents of the Stream Deck buttons.
         /// </summary>
-        public void ClearPanel()
+        public void ClearButtons()
         {
             for (int i = 0; i < this.ButtonCount; i++)
             {
-                // TODO: Need to replace this with device-specific logic
-                // since not every device is 96x96.
-                this.SetKey(i, DeviceConstants.XLDefaultBlackButton);
+                this.SetKey(i, ImageHelpers.CreateBlankImage(this.ButtonResolution, Color.Black));
             }
         }
 
@@ -172,57 +216,121 @@ namespace DeckSurf.SDK.Models
                     {
                         byte[] imageBuffer = File.ReadAllBytes(button.ButtonImagePath);
 
-                        // TODO: Need to make sure that I am using device-agnostic button sizes.
-                        imageBuffer = ImageHelpers.ResizeImage(imageBuffer, DeviceConstants.XLButtonSize, DeviceConstants.XLButtonSize);
+                        imageBuffer = ImageHelpers.ResizeImage(imageBuffer, this.ButtonResolution, this.ButtonResolution, this.IsButtonImageFlipRequired);
                         this.SetKey(button.ButtonIndex, imageBuffer);
                     }
                 }
             }
         }
 
-
         /// <summary>
         /// Sets the content of a key on a Stream Deck device.
         /// </summary>
-        /// <param name="keyId">Numberic ID of the key that needs to be set.</param>
+        /// <param name="keyId">Numeric ID of the key that needs to be set.</param>
         /// <param name="image">Binary content (JPEG) of the image that needs to be set on the key. The image will be resized to match the expectations of the connected device.</param>
         /// <returns>True if succesful, false if not.</returns>
         public bool SetKey(int keyId, byte[] image)
         {
-            var content = image ?? DeviceConstants.XLDefaultBlackButton;
+            var iteration = 0;
+            var remainingBytes = image.Length;
+
+            using var stream = this.Open();
+            while (remainingBytes > 0)
+            {
+                var sliceLength = Math.Min(remainingBytes, ImageReportPayloadLength);
+                var bytesSent = iteration * ImageReportPayloadLength;
+
+                byte finalizer = sliceLength == remainingBytes ? (byte)1 : (byte)0;
+
+                var binaryLength = DataHelpers.GetLittleEndianBytesFromInt(sliceLength);
+                var binaryIteration = DataHelpers.GetLittleEndianBytesFromInt(iteration);
+
+                // TODO: This is different for different device classes, so I will need
+                // to make sure that I adjust the header format.
+                byte[] header =
+                [
+                    0x02,
+                    0x07,
+                    (byte)keyId,
+                    finalizer,
+                    binaryLength[0],
+                    binaryLength[1],
+                    binaryIteration[0],
+                    binaryIteration[1]
+                ];
+
+                var payload = header.Concat(new ArraySegment<byte>(image, bytesSent, sliceLength)).ToArray();
+                var padding = new byte[ImageReportLength - payload.Length];
+
+                var finalPayload = payload.Concat(padding).ToArray();
+
+                stream.Write(finalPayload);
+
+                remainingBytes -= sliceLength;
+                iteration++;
+            }
+
+            return true;
+        }
+
+        /// <summary>
+        /// Sets the screen image for a connected Stream Deck device.
+        /// </summary>
+        /// <remarks>Currently only supported for the Stream Deck Plus.</remarks>
+        /// <param name="image">Binary content (JPEG) of the image that needs to be set on the screen. The image will be resized to match the expectations of the connected device.</param>
+        /// <param name="offset">Offset from the left where the image needs to be set. Set to zero if setting the full image.</param>
+        /// <param name="width">Image height.</param>
+        /// <param name="height">Image width.</param>
+        /// <returns>True if succesful, false if not.</returns>
+        public bool SetScreen(byte[] image, int offset, int width, int height)
+        {
+            byte[] binaryOffset = DataHelpers.GetLittleEndianBytesFromInt(offset);
+            byte[] binaryWidth = DataHelpers.GetLittleEndianBytesFromInt(width);
+            byte[] binaryHeight = DataHelpers.GetLittleEndianBytesFromInt(height);
 
             var iteration = 0;
-            var remainingBytes = content.Length;
+            var remainingBytes = image.Length;
 
-            using (var stream = this.Open())
+            using var stream = this.Open();
+            while (remainingBytes > 0)
             {
-                while (remainingBytes > 0)
-                {
-                    var sliceLength = Math.Min(remainingBytes, ImageReportPayloadLength);
-                    var bytesSent = iteration * ImageReportPayloadLength;
+                var sliceLength = Math.Min(remainingBytes, ImageReportScreenPayloadLength);
+                var bytesSent = iteration * ImageReportScreenPayloadLength;
 
-                    byte finalizer = sliceLength == remainingBytes ? (byte)1 : (byte)0;
+                byte isLastChunk = sliceLength == remainingBytes ? (byte)1 : (byte)0;
 
-                    // These components are nothing else but UInt16 low-endian
-                    // representations of the length of the image payload, and iteration.
-                    var bitmaskedLength = (byte)(sliceLength & 0xFF);
-                    var shiftedLength = (byte)(sliceLength >> ImageReportHeaderLength);
-                    var bitmaskedIteration = (byte)(iteration & 0xFF);
-                    var shiftedIteration = (byte)(iteration >> ImageReportHeaderLength);
+                var binaryLength = DataHelpers.GetLittleEndianBytesFromInt(sliceLength);
+                var binaryIteration = DataHelpers.GetLittleEndianBytesFromInt(iteration);
 
-                    // TODO: This is different for different device classes, so I will need
-                    // to make sure that I adjust the header format.
-                    byte[] header = new byte[] { 0x02, 0x07, (byte)keyId, finalizer, bitmaskedLength, shiftedLength, bitmaskedIteration, shiftedIteration };
-                    var payload = header.Concat(new ArraySegment<byte>(content, bytesSent, sliceLength)).ToArray();
-                    var padding = new byte[ImageReportLength - payload.Length];
+                byte[] header =
+                [
+                    0x02,
+                    0x0C,
+                    binaryOffset[0],
+                    binaryOffset[1],
+                    0x00,
+                    0x00,
+                    binaryWidth[0],
+                    binaryWidth[1],
+                    binaryHeight[0],
+                    binaryHeight[1],
+                    isLastChunk,
+                    binaryIteration[0],
+                    binaryIteration[1],
+                    binaryLength[0],
+                    binaryLength[1],
+                    0x00
+                ];
 
-                    var finalPayload = payload.Concat(padding).ToArray();
+                var payload = header.Concat(new ArraySegment<byte>(image, bytesSent, sliceLength)).ToArray();
+                var padding = new byte[ImageReportLength - payload.Length];
 
-                    stream.Write(finalPayload);
+                var finalPayload = payload.Concat(padding).ToArray();
 
-                    remainingBytes -= sliceLength;
-                    iteration++;
-                }
+                stream.Write(finalPayload);
+
+                remainingBytes -= sliceLength;
+                iteration++;
             }
 
             return true;
@@ -230,20 +338,101 @@ namespace DeckSurf.SDK.Models
 
         private void KeyPressCallback(IAsyncResult result)
         {
+            var buttonMapOffset = 4;
+
             int bytesRead = this.UnderlyingInputStream.EndRead(result);
 
-            var buttonData = new ArraySegment<byte>(this.keyPressBuffer, ButtonPressHeaderOffset, ButtonCount).ToArray();
-            var pressedButton = Array.IndexOf(buttonData, (byte)1);
-            var buttonKind = pressedButton != -1 ? ButtonEventKind.DOWN : ButtonEventKind.UP;
+            // Let's grab the first two bytes to understand the type of button we're dealing with.
+            // They can be:
+            //    0x01 0x00 - Button
+            //    0x01 0x02 - Touch screen
+            //    0x01 0x03 - Knob
+            var header = new ArraySegment<byte>(this.keyPressBuffer, 0, 2).ToArray();
+            var buttonKind = this.GetButtonKind(header);
+            var isKnobRotated = false;
+            var knobRotationDirection = KnobRotationDirection.None;
+            var buttonCount = DataHelpers.GetIntFromLittleEndianBytes(new ArraySegment<byte>(this.keyPressBuffer, 2, 2).ToArray());
+
+            // If this was not a touch screen, we should provide
+            // dummy coordinates.
+            Point touchPoint = new() { X = -1, Y = -1 };
+
+            if (buttonKind == ButtonKind.Screen)
+            {
+                var xCoord = new ArraySegment<byte>(this.keyPressBuffer, 6, 2).ToArray();
+                var yCoord = new ArraySegment<byte>(this.keyPressBuffer, 8, 2).ToArray();
+
+                touchPoint = new Point() { X = DataHelpers.GetIntFromLittleEndianBytes(xCoord), Y = DataHelpers.GetIntFromLittleEndianBytes(yCoord) };
+            }
+
+            // For whatever reason, the number of knobs is reported as 5, even though
+            // there are only 4 on the Stream Deck Plus. Because that's the only device
+            // where that value is used today, let's make sure that we decrement by 1.
+            // Also, for the knob, the header is 5 bytes long, because the fifth
+            // byte tells us whether the knob is rotated or not.
+            if (buttonKind == ButtonKind.Knob)
+            {
+                buttonCount -= 1;
+                buttonMapOffset += 1;
+            }
+
+            var buttonData = new ArraySegment<byte>(this.keyPressBuffer, buttonMapOffset, buttonCount).ToArray();
+
+            int pressedButton = -1;
+
+            if (buttonKind == ButtonKind.Button || buttonKind == ButtonKind.Screen)
+            {
+                pressedButton = Array.IndexOf(buttonData, (byte)0x01);
+            }
+            else
+            {
+                isKnobRotated = this.keyPressBuffer[4] != (byte)0x00;
+
+                pressedButton = Array.IndexOf(buttonData, (byte)0x01);
+                if (pressedButton == -1)
+                {
+                    pressedButton = Array.IndexOf(buttonData, (byte)0xFF);
+
+                    if (isKnobRotated)
+                    {
+                        knobRotationDirection = KnobRotationDirection.Left;
+                    }
+                }
+                else
+                {
+                    if (isKnobRotated)
+                    {
+                        knobRotationDirection = KnobRotationDirection.Right;
+                    }
+                }
+            }
+
+            var eventKind = pressedButton != -1 ? ButtonEventKind.DOWN : ButtonEventKind.UP;
 
             if (this.OnButtonPress != null)
             {
-                this.OnButtonPress(this.UnderlyingDevice, new ButtonPressEventArgs(pressedButton, buttonKind));
+                this.OnButtonPress(this.UnderlyingDevice, new ButtonPressEventArgs(pressedButton, eventKind, buttonKind, touchPoint, isKnobRotated, knobRotationDirection));
             }
 
             Array.Clear(this.keyPressBuffer, 0, this.keyPressBuffer.Length);
 
             this.UnderlyingInputStream.BeginRead(this.keyPressBuffer, 0, this.keyPressBuffer.Length, this.KeyPressCallback, null);
+        }
+
+        private ButtonKind GetButtonKind(byte[] identifier)
+        {
+            if (identifier.Length != 2)
+            {
+                return ButtonKind.Unknown;
+            }
+
+            return (identifier[0], identifier[1]) switch
+            {
+                (0x01, 0x00) => ButtonKind.Button,
+                (0x01, 0x02) => ButtonKind.Screen,
+                (0x01, 0x03) => ButtonKind.Knob,
+                _ => ButtonKind.Unknown,
+            };
         }
     }
 }
