@@ -5,15 +5,15 @@
 namespace DeckSurf.SDK.Models.Devices
 {
     /// <summary>
-    /// Implementation for a Stream Deck XL connected device.
+    /// Implementation for a Stream Deck Original V2 connected device.
     /// </summary>
-    public class StreamDeckXL(int vid, int pid, string path, string name) : ConnectedDevice(vid, pid, path, name)
+    internal class StreamDeckOriginalV2(int vid, int pid, string path, string name) : ConnectedDevice(vid, pid, path, name)
     {
         /// <inheritdoc/>
-        public override DeviceModel Model => DeviceModel.XL;
+        public override DeviceModel Model => DeviceModel.ORIGINAL_V2;
 
         /// <inheritdoc/>
-        public override int ButtonCount => 32;
+        public override int ButtonCount => 15;
 
         /// <inheritdoc/>
         public override bool IsButtonImageFlipRequired => true;
@@ -25,13 +25,13 @@ namespace DeckSurf.SDK.Models.Devices
         public override bool IsKnobSupported => false;
 
         /// <inheritdoc/>
-        public override int ButtonResolution => 96;
+        public override int ButtonResolution => 72;
 
         /// <inheritdoc/>
-        public override int ButtonColumns => 8;
+        public override int ButtonColumns => 5;
 
         /// <inheritdoc/>
-        public override int ButtonRows => 4;
+        public override int ButtonRows => 3;
 
         /// <inheritdoc/>
         public override int ScreenWidth => -1;
