@@ -1,4 +1,8 @@
-﻿using System;
+﻿// Copyright (c) Den Delimarsky
+// Den Delimarsky licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
 using System.Runtime.InteropServices;
 using DeckSurf.SDK.Models;
 
